requirements.txt
openai>=1.12.0
google-generativeai>=0.3.0
tavily-python>=0.3.0
httpx>=0.25.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.18.0
numpy>=1.24.0
scipy>=1.11.0